missed some casts

+++ src/server/scripts/EasternKingdoms/BlackrockMountain/BlackrockDepths/boss_gorosh_the_dervish.cpp
@@ -52,9 +52,9 @@ public:
         void EnterCombat(Unit* /*who*/) override
         {
             _EnterCombat();
-            events.ScheduleEvent(SPELL_WHIRLWIND, 0.2 * TIMER_WHIRLWIND);
-            events.ScheduleEvent(SPELL_MORTALSTRIKE, 0.2 * TIMER_MORTAL);
-            events.ScheduleEvent(SPELL_BLOODLUST, 0.2 * TIMER_BLOODLUST);
+            events.ScheduleEvent(SPELL_WHIRLWIND, 0.2 * (int) TIMER_WHIRLWIND);
+            events.ScheduleEvent(SPELL_MORTALSTRIKE, 0.2 * (int) TIMER_MORTAL);
+            events.ScheduleEvent(SPELL_BLOODLUST, 0.2 * (int) TIMER_BLOODLUST);
         }
 
         void UpdateAI(uint32 diff) override
